requirements-prod.txt
# Production requirements for StoxChai
# Core dependencies
streamlit==1.32.0
yfinance==0.2.54
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0

# Data processing and ML
scikit-learn==1.3.2
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.30.2
torch==2.1.2
tiktoken==0.5.2

# Web scraping and APIs
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8

# Database and caching
psycopg2-binary==2.9.7
redis==4.6.0
sqlalchemy==2.0.23

# AI and LangChain
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.60
langchain-ollama==0.3.3
ollama==0.4.8

# Vector database
faiss-cpu==1.7.4
chromadb==0.4.15

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Security
cryptography==41.0.7
passlib==1.7.4
python-jose==3.3.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2
click==8.1.7

# Performance monitoring
psutil==5.9.6
memory-profiler==0.61.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Production server
gunicorn==21.2.0
uvicorn==0.24.0